refactor(controller) : 폴더 구조 및 파일명 수정
- 기존 RaceController 클래스의 이름을 Application으로 수정
- RaceGame을 도메인에서 controller 패키지로 이동
